FISH-9513 Moved obfuscation to a util class.
As per Kalins request, I have moved the process of obfuscation to another class and did some name changes.
