+++ src/main/java/fish/payara/extras/diagnostics/collection/collectors/DomainXmlCollector.java
@@ -1,64 +1,12 @@
-/*
- * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS HEADER.
- *
- * Copyright (c) 2023-2024 Payara Foundation and/or its affiliates. All rights reserved.
- *
- * The contents of this file are subject to the terms of either the GNU
- * General Public License Version 2 only ("GPL") or the Common Development
- * and Distribution License("CDDL") (collectively, the "License").  You
- * may not use this file except in compliance with the License.  You can
- * obtain a copy of the License at
- * https://glassfish.dev.java.net/public/CDDL+GPL_1_1.html
- * or packager/legal/LICENSE.txt.  See the License for the specific
- * language governing permissions and limitations under the License.
- *
- * When distributing the software, include this License Header Notice in each
- * file and include the License file at packager/legal/LICENSE.txt.
- *
- * GPL Classpath Exception:
- * Oracle designates this particular file as subject to the "Classpath"
- * exception as provided by Oracle in the GPL Version 2 section of the License
- * file that accompanied this code.
- *
- * Modifications:
- * If applicable, add the following below the License Header, with the fields
- * enclosed by brackets [] replaced by your own identifying information:
- * "Portions Copyright [year] [name of copyright owner]"
- *
- * Contributor(s):
- * If you wish your version of this file to be governed by only the CDDL or
- * only the GPL Version 2, indicate your decision by adding "[Contributor]
- * elects to include this software in this distribution under the [CDDL or GPL
- * Version 2] license."  If you don't indicate a single choice of license, a
- * recipient has the option to distribute your version of this file under
- * either the CDDL, the GPL Version 2 or to extend the choice of license to
- * its licensees as provided above.  However, if you add GPL Version 2 code
- * and therefore, elected the GPL Version 2 license, then the option applies
- * only if the new code is made subject to such option by the copyright
- * holder.
- */
-
 package fish.payara.extras.diagnostics.collection.collectors;
 
+import fish.payara.extras.diagnostics.util.ObfuscateDomainXml;
 import fish.payara.extras.diagnostics.util.ParamConstants;
-import org.w3c.dom.Document;
-import org.w3c.dom.Element;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
 
-import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Map;
 import java.util.logging.Logger;
-
-import javax.xml.parsers.DocumentBuilder;
-import javax.xml.parsers.DocumentBuilderFactory;
-import javax.xml.transform.OutputKeys;
-import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerFactory;
-import javax.xml.transform.dom.DOMSource;
-import javax.xml.transform.stream.StreamResult;
 
 public class DomainXmlCollector extends FileCollector {
 
@@ -66,10 +14,8 @@ public class DomainXmlCollector extends FileCollector {
     private String dirSuffix;
     private Logger LOGGER = Logger.getLogger(DomainXmlCollector.class.getName());
     private boolean obfuscateDomainXml;
-    private final String PASSWORD_CHANGE = "PASSWORD_HIDDEN";
-    private final String PASSWORD_KEYWORD = "password";
     private final int COLLECTED_OKAY = 0;
-
+    private final ObfuscateDomainXml obfuscate = new ObfuscateDomainXml();
 
     public DomainXmlCollector(Path path, boolean obfuscateDomainXml) {
         this.path = path;
@@ -83,7 +29,6 @@ public class DomainXmlCollector extends FileCollector {
         this.obfuscateDomainXml = obfuscateDomainXml;
     }
 
-
     @Override
     public int collect() {
         int domainXmlCollected = COLLECTED_OKAY;
@@ -96,7 +41,7 @@ public class DomainXmlCollector extends FileCollector {
                 LOGGER.info("Collecting domain.xml from " + (getInstanceName() != null ? getInstanceName() : "server"));
                 domainXmlCollected = super.collect();
                 if (domainXmlCollected == COLLECTED_OKAY && obfuscateDomainXml) {
-                    modifyXMLFile(resolveDestinationFile().toFile());
+                    obfuscate.obfuscateDomainXml(resolveDestinationFile().toFile());
                 }
             }
         }
@@ -113,57 +58,4 @@ public class DomainXmlCollector extends FileCollector {
         }
         return null;
     }
-
-    private void modifyXMLFile(File xmlFile) {
-        LOGGER.info("Modifying " + xmlFile.getAbsolutePath());
-        modifyXMLPassword(xmlFile);
-    }
-
-    private void modifyXMLPassword(File xmlFile) {
-        try {
-            DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
-            DocumentBuilder dBuilder = dbFactory.newDocumentBuilder();
-            Document doc = dBuilder.parse(xmlFile);
-            doc.getDocumentElement().normalize();
-            traverseNodes(doc.getDocumentElement());
-
-
-            TransformerFactory transformerFactory = TransformerFactory.newInstance();
-            Transformer transformer = transformerFactory.newTransformer();
-            transformer.setOutputProperty(OutputKeys.INDENT,"yes");
-            DOMSource source = new DOMSource(doc);
-            StreamResult result = new StreamResult(xmlFile);
-            transformer.transform(source, result);
-
-            LOGGER.info("Successfully modified " + xmlFile.getAbsolutePath());
-        } catch (Exception e) {
-            LOGGER.severe("Error Modifying XML : " + e.getMessage());
-            e.printStackTrace();
-        }
-    }
-
-    private void traverseNodes(Node node) {
-
-        if (node.getNodeType() == Node.ELEMENT_NODE) {
-            Element tempNode = (Element) node;
-            boolean hasPasswordAttribute = tempNode.hasAttribute(PASSWORD_KEYWORD);
-            if (hasPasswordAttribute) {
-                tempNode.setAttribute(PASSWORD_KEYWORD, PASSWORD_CHANGE);
-            }
-            String nameAttribute = tempNode.getAttribute("name");
-            boolean hasValueAttribute = tempNode.hasAttribute("value");
-            if (PASSWORD_KEYWORD.equalsIgnoreCase(nameAttribute)) {
-                if (hasValueAttribute) {
-                    tempNode.setAttribute("value", PASSWORD_CHANGE);
-                }
-            }
-
-            NodeList childNodes = node.getChildNodes();
-            for (int i = 0; i < childNodes.getLength(); i++) {
-                traverseNodes(childNodes.item(i));
-            }
-        }
-    }
-
-
 }
